+++ threedigrid_builder/tests/test_gridadmin.py
@@ -131,6 +131,7 @@ def h5_out():
         ("dmax", (4,), "float64"),  # added
         ("s1d", (4,), "float64"),  # added
         ("embedded_in", (4,), "int32"),  # added
+        ("boundary_type", (4,), "int32"),  # added
     ],
 )
 def test_write_nodes(h5_out, dataset, shape, dtype):
@@ -168,6 +169,8 @@ def test_write_nodes(h5_out, dataset, shape, dtype):
         ("dmax", (3,), "float64"),  # added
         ("s1d", (3,), "float64"),  # added
         ("embedded_in", (3,), "int32"),  # added
+        ("boundary_id", (3,), "int32"),  # added
+        ("boundary_type", (3,), "int32"),  # added
     ],
 )
 def test_write_nodes_embedded(h5_out, dataset, shape, dtype):