+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog of threedigrid-builder
 0.3.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Added 1D boundary conditions.
 
 
 0.3.1 (2021-08-16)